Går bort i fra spring autowire og newer opp instanset istedenfor
