+++ src/test/java/no/nav/foreldrepenger/mottak/http/RestClientConfigurationTest.java
@@ -8,13 +8,10 @@ import java.util.Map;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import com.nimbusds.oauth2.sdk.auth.ClientAuthenticationMethod;
 
-import no.nav.foreldrepenger.mottak.config.TokenConfigMatcherConfiguration;
 import no.nav.security.token.support.client.core.ClientAuthenticationProperties;
 import no.nav.security.token.support.client.core.ClientProperties;
 import no.nav.security.token.support.client.core.OAuth2GrantType;
@@ -22,12 +19,9 @@ import no.nav.security.token.support.client.spring.ClientConfigurationProperties
 import no.nav.security.token.support.client.spring.oauth2.ClientConfigurationPropertiesMatcher;
 
 @ExtendWith(SpringExtension.class)
-@Import(value = TokenConfigMatcherConfiguration.class)
 class RestClientConfigurationTest {
 
-    @Autowired
-    private ClientConfigurationPropertiesMatcher matcher;
-
+    private static ClientConfigurationPropertiesMatcher matcher = new RestClientConfiguration().tokenxClientConfigMatcher();
     private static ClientConfigurationProperties properties;
 
     @BeforeAll
